optimized sphere creation, auto relative mouse control in level 7

+++ WinRTDX/Attachable.h
@@ -9,7 +9,7 @@ namespace Dx::Attachables
 	{
 	public:
 
-		Attachable(uint16_t key, std::shared_ptr<Dx::Graphics> graphics) :
+		Attachable(int key, std::shared_ptr<Dx::Graphics> graphics) :
 			m_graphics(graphics), 
 			m_device(graphics->Device()),
 			m_context(graphics->Context()),
@@ -23,6 +23,6 @@ namespace Dx::Attachables
 		std::shared_ptr<Dx::Graphics>					m_graphics;
 		com_ptr<ID3D11Device3>							m_device;
 		com_ptr<ID3D11DeviceContext4>					m_context;
-		uint16_t												m_key;
+		int													m_key;
 	};
 }
